added a bunch of keyframe stuff

+++ src/cs3500/animator/model/keyframe/ReadOnlyIKeyFrame.java
@@ -11,11 +11,11 @@ public interface ReadOnlyIKeyFrame {
 
   double getH();
 
-  double getR();
+  int getR();
 
-  double getG();
+  int getG();
 
-  double getB();
+  int getB();
 
-  double getT();
+  int getT();
 }
